% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rray3-package.R
\docType{package}
\name{rray3-package}
\alias{rray3}
\alias{rray3-package}
\title{rray3: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/rray3}
  \item Report bugs at \url{https://github.com/DavisVaughan/rray3/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
